% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{cam_colors}
\alias{cam_colors}
\title{Cambridge MA Color Palettes}
\format{
An object of class \code{character} of length 23.
}
\usage{
cam_colors
}
\description{
Named vector of all Cambridge-inspired colors
}
\examples{
cam_colors["red_maple_flower"]
cam_colors[c("snow", "granite")]
}
\keyword{datasets}
