% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cambridgecolors.R
\name{cam_cols}
\alias{cam_cols}
\title{Get n Cambridge colors}
\usage{
cam_cols(n, palette = "spring", reverse = FALSE)
}
\arguments{
\item{n}{Number of colors needed}

\item{palette}{Palette name: "spring", "summer", "autumn", "winter", or "river"}

\item{reverse}{Logical, reverse color order if TRUE}
}
\value{
Character vector of n hex colors
}
\description{
Base R helper to get n colors from a Cambridge palette
}
\examples{
cam_cols(3, "spring")
barplot(1:5, col = cam_cols(5, "autumn"))
}
