% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{list_palettes}
\alias{list_palettes}
\title{List available palettes}
\usage{
list_palettes()
}
\value{
A named list with palette names grouped by type
}
\description{
Returns a list of all available palette names grouped by type
}
\examples{
list_palettes()
}
