% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_cam_demo}
\alias{plot_cam_demo}
\title{Display Cambridge color palettes}
\usage{
plot_cam_demo()
}
\value{
NULL (displays plot)
}
\description{
Shows a swatch panel of all available Cambridge palettes
}
\examples{
plot_cam_demo()
}
