% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorblind.R
\name{simulate_cvd}
\alias{simulate_cvd}
\title{Simulate color vision deficiency}
\usage{
simulate_cvd(colors, type = c("deuteranopia", "protanopia", "tritanopia"))
}
\arguments{
\item{colors}{Character vector of hex colors}

\item{type}{Type of CVD: "deuteranopia", "protanopia", or "tritanopia"}
}
\value{
Character vector of transformed hex colors
}
\description{
Transforms colors to simulate how they appear to people with
different types of color vision deficiency.
}
\examples{
# Simulate deuteranopia
spring_cols <- cam_palette("spring")
simulate_cvd(spring_cols, "deuteranopia")
}
