% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cambridgecolors.R
\name{cam_pal}
\alias{cam_pal}
\title{Cambridge palette function generator}
\usage{
cam_pal(palette = "spring", reverse = FALSE)
}
\arguments{
\item{palette}{Palette name: "spring", "summer", "autumn", "winter", or "river"}

\item{reverse}{Logical, reverse color order if TRUE}
}
\value{
Function that takes n and returns n hex colors
}
\description{
Returns a function that generates n colors from a palette
}
\examples{
spring_pal <- cam_pal("spring")
spring_pal(3)
cam_pal("autumn")(7)
}
